code clean up: correcting, removed unused, moving to own file
